Drop old RelayList and MullvadRpc
