+++ ios/MullvadVPN/AnyEncodable.swift
@@ -1,0 +1,28 @@
+//
+//  AnyEncodable.swift
+//  MullvadVPN
+//
+//  Created by pronebird on 21/07/2020.
+//  Copyright © 2020 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+
+extension Encodable {
+    fileprivate func encode(to container: inout SingleValueEncodingContainer) throws {
+        try container.encode(self)
+    }
+}
+
+struct AnyEncodable: Encodable {
+    let value: Encodable
+
+    init(_ value: Encodable) {
+        self.value = value
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try value.encode(to: &container)
+    }
+}
