+++ ios/MullvadVPN/DisplayChainedError.swift
@@ -36,31 +36,6 @@ extension RestError: DisplayChainedError {
     }
 }
 
-extension MullvadRpc.Error: DisplayChainedError {
-    var errorChainDescription: String? {
-        switch self {
-        case .network(let urlError):
-            return urlError.localizedDescription
-
-        case .server(let serverError):
-            if let knownErrorDescription = serverError.errorDescription {
-                return knownErrorDescription
-            } else {
-                return String(
-                    format: NSLocalizedString("Server error: %@", comment: ""),
-                    serverError.message
-                )
-            }
-
-        case .encoding:
-            return NSLocalizedString("Server request encoding error", comment: "")
-
-        case .decoding:
-            return NSLocalizedString("Server response decoding error", comment: "")
-        }
-    }
-}
-
 extension TunnelManager.Error: DisplayChainedError {
     var errorChainDescription: String? {
         switch self {
@@ -97,22 +72,22 @@ extension TunnelManager.Error: DisplayChainedError {
         case .removeTunnelSettings(_):
             return NSLocalizedString("Failed to remove tunnel settings", comment: "")
 
-        case .pushWireguardKey(let rpcError):
-            let reason = rpcError.errorChainDescription ?? ""
+        case .pushWireguardKey(let restError):
+            let reason = restError.errorChainDescription ?? ""
             var message = String(format: NSLocalizedString("Failed to send the WireGuard key to server: %@", comment: ""), reason)
 
-            if case .server(let serverError) = rpcError, serverError.code == .tooManyWireguardKeys {
+            if case .server(.keyLimitReached) = restError {
                 message.append("\n\n")
                 message.append(NSLocalizedString("Remove unused WireGuard keys and try again", comment: ""))
             }
 
             return message
 
-        case .replaceWireguardKey(let rpcError):
-            let reason = rpcError.errorChainDescription ?? ""
+        case .replaceWireguardKey(let restError):
+            let reason = restError.errorChainDescription ?? ""
             var message = String(format: NSLocalizedString("Failed to replace the WireGuard key on server: %@", comment: ""), reason)
 
-            if case .server(let serverError) = rpcError, serverError.code == .tooManyWireguardKeys {
+            if case .server(.keyLimitReached) = restError {
                 message.append("\n\n")
                 message.append(NSLocalizedString("Remove unused WireGuard keys and try again", comment: ""))
             }
@@ -123,8 +98,8 @@ extension TunnelManager.Error: DisplayChainedError {
             // This error is never displayed anywhere
             return nil
 
-        case .verifyWireguardKey(let rpcError):
-            let reason = rpcError.errorChainDescription ?? ""
+        case .verifyWireguardKey(let restError):
+            let reason = restError.errorChainDescription ?? ""
 
             return String(format: NSLocalizedString("Failed to verify the WireGuard key on server: %@", comment: ""), reason)
 
@@ -157,8 +132,8 @@ extension AppStorePaymentManager.Error: DisplayChainedError {
         case .readReceipt(let readReceiptError):
             return String(format: NSLocalizedString("Cannot read the receipt: %@", comment: ""), readReceiptError.errorChainDescription ?? "")
 
-        case .sendReceipt(let rpcError):
-            let reason = rpcError.errorChainDescription ?? ""
+        case .sendReceipt(let restError):
+            let reason = restError.errorChainDescription ?? ""
 
             return String(format: NSLocalizedString(#"Failed to send the receipt to server: %@\n\nPlease retry by using the "Restore purchases" button."#, comment: ""), reason)
